LayoutTests/http/tests/websocket/tests/hybi/bad-sub-protocol-empty.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test WebSocket bad sub-protocol names (empty).");

if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

// Fails if protocol is an empty string.
shouldThrow('new WebSocket("ws://127.0.0.1:8880/simple", "")');

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
